refactor the logic
