+++ client/a8c-for-agencies/sections/reports/hooks/use-fetch-report-by-id.tsx
@@ -2,12 +2,12 @@ import { useQuery } from '@tanstack/react-query';
 import wpcom from 'calypso/lib/wp';
 import { useSelector } from 'calypso/state';
 import { getActiveAgencyId } from 'calypso/state/a8c-for-agencies/agency/selectors';
-import type { Report } from '../types';
+import type { Report, ReportError } from '../types';
 
 export default function useFetchReportById( reportId: number | null ) {
 	const agencyId = useSelector( getActiveAgencyId );
 
-	return useQuery< Report, Error >( {
+	return useQuery< Report, ReportError >( {
 		queryKey: [ 'a4a-report', agencyId, reportId ],
 		queryFn: async () => {
 			return wpcom.req.get( {
@@ -17,7 +17,10 @@ export default function useFetchReportById( reportId: number | null ) {
 		},
 		enabled: !! agencyId && !! reportId,
 		refetchOnWindowFocus: false,
-		retry: ( failureCount ) => {
+		retry: ( failureCount, error ) => {
+			if ( error.data.status === 404 ) {
+				return false;
+			}
 			return failureCount < 3;
 		},
 	} );
