+++ client/a8c-for-agencies/sections/reports/primary/build-report/style.scss
@@ -127,3 +127,9 @@
 	display: flex;
 	gap: 8px;
 }
+
+.build-report__step-note {
+	@include body-small;
+	color: var(--color-neutral-50);
+	margin-block: 0.5rem -0.5rem;
+}
